Up Rev'd content to track changes that were actually made in the wiki...
Noted that the editing in the Wiki was pretty easy... however, as a practice (that will appear in the soon to be released documentation) the editing is intended to be done in this GitHub repo file so as to enable diff's and other functions.  Evaluating the best methods and practices for use of GitHub for documents of this type is an ongoing aspect of this research initiative.