replaced grant/revoke access, added protect
Replaced grant/revoke access with simply "manage", intended to include managing not only "access" rights but more generally to grant, monitor, modify (if possible) and revoke authorizations to protected resources of the owner, in general.  

Also therefore added the concept of "protect" to signify the right or obligation to secure and exclude others from access, use or benefit of the resource.